feat(interface): repo capabilities + ErrorCode enum migration + test split
Three deliverables from interface-api-v1-eval P1:

1. **list_repos / add_repo / remove_repo** capabilities. Wraps Config's
   existing global_index_repos / project_index_repos accessors and the
   on-disk read-modify-write of $XLINGS_HOME/.xlings.json's index_repos
   array. add_repo is upsert (matches the CLI's --index-repo behavior).
   list_repos emits dataKind:repo_list with each entry tagged
   scope:"global"|"project". Capability count: 15 → 18.

2. **ErrorCode int → enum string**. ErrorEvent.code was an opaque int
   (every error came out as code:1 for capabilities, 404/500 only at
   the dispatcher). Replaced with an ErrorCode enum class + a stable
   wire alphabet — E_INVALID_INPUT / E_NOT_FOUND / E_NETWORK /
   E_DISK_FULL / E_PERMISSION / E_CANCELLED / E_INTERNAL — written as
   a JSON string. ErrorEvent also gains an optional `hint` field for
   remediation text. Clients can now match on a stable code instead
   of grepping the message.

3. **Tests split**: pulled the InterfaceProtocol suite (now 17 tests)
   out of the 4000-line tests/unit/test_main.cpp into its own
   tests/unit/test_interface_protocol.cpp. xmake's add_files glob
   already picks up the new file, so no build config change.
   New tests cover the full repo lifecycle and the enum-string
   error format.

   Note: file-local helpers in capabilities.cppm intentionally
   non-inline. inline + module + libstdc++ static link causes weak
   symbols touching nlohmann::json to demand
   std::shared_ptr<output_adapter_protocol<char>>'s copy ctor, which
   isn't resolvable in static-link mode. Inline comment in the file
   warns the next person.

186 tests run, 182 pass, 4 skipped (require XIM_PKGINDEX_DIR — pre-existing).
